Add the source files `main.rs`, `rulesets.rs`, `tm.rs`, and `turing_machine_memory.rs`.
`rulesets.rs` handles the creation, storage and modification of rulesets.
A *rule* is a mapping (current_state, current_symbol) => (new_state, new_symbol, move_direction) and
a *ruleset* is a set of mappings with all possible tuples (current_state, current_symbols).
With 4 possible states and 3 possible symbols the ruleset has 4 * 3 = 12 entries.

`turing_machine_memory.rs` holds the memory. It is implemented for 1d, 2d and 3d memory.
It also handles the "tapehead" movement.

`tm.rs` is the actual turing machine. It is generic in terms of the memory being used.